Unpublish endpoint (#49)
* Add unpublish endpoint to OpenAPI spec.
Add unpublish method to internal Discover client and test helpers.

* Working on unpublish route.

* Testing unpublish.

* Testing unpublish.

* Test all permutations of publish status during publish.

* Fix scoping of table driven test.

+++ internal/api/publishing/manifests.go
@@ -33,8 +33,8 @@ const ManifestSchemaVersion = "http://schema.org/version/3.7/"
 const ManifestPennsieveSchemaVersion = "5.0"
 
 type ManifestV5 struct {
-	PennsieveDatasetID int64                  `json:"pennsieveDatasetId"`
-	Version            int64                  `json:"version"`
+	PennsieveDatasetID int                    `json:"pennsieveDatasetId"`
+	Version            int                    `json:"version"`
 	Revision           int                    `json:"revision,omitempty"`
 	Name               string                 `json:"name,omitempty"`
 	Description        string                 `json:"description"`
@@ -78,7 +78,7 @@ func (m ManifestV5) TotalSize() int64 {
 	return totalSize
 }
 
-func ManifestS3Key(publishedDatasetID int64) string {
+func ManifestS3Key(publishedDatasetID int) string {
 	return fmt.Sprintf("%d/%s", publishedDatasetID, ManifestFileName)
 }
 
@@ -115,12 +115,12 @@ func NewManifestBuilder() *ManifestBuilder {
 	}}
 }
 
-func (b *ManifestBuilder) WithPennsieveDatasetID(id int64) *ManifestBuilder {
+func (b *ManifestBuilder) WithPennsieveDatasetID(id int) *ManifestBuilder {
 	b.m.PennsieveDatasetID = id
 	return b
 }
 
-func (b *ManifestBuilder) WithVersion(version int64) *ManifestBuilder {
+func (b *ManifestBuilder) WithVersion(version int) *ManifestBuilder {
 	b.m.Version = version
 	return b
 }
